Made entities die-ables

+++ src/main/java/me/maxiiiiii/skyblockdragons/damage/events/EntityDamageEvent.java
@@ -14,7 +14,7 @@ public class EntityDamageEvent extends Event {
     private final EntityDamage damage;
     private final EntitySD attacker;
     private final EntitySD victim;
-    private final double finalDamage;
+    private final long finalDamage;
 
     public EntityDamageEvent(EntityDamage damage) {
         this.damage = damage;
